% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_CL.R
\name{convert_BASF_CL}
\alias{convert_BASF_CL}
\title{convert_BASF_CL}
\usage{
convert_BASF_CL(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convert_BASF_CL
}
\examples{
 domainName<-'CL'
 domainData<-loadDomain(domainName)
 SEND<-convertCL(domainName, domainData)

}
