% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_FM.R
\name{convert_BASF_FM}
\alias{convert_BASF_FM}
\title{convert_BASF_FM}
\usage{
convert_BASF_FM(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The dataframe containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe, use BASF Provantis Export files
}
\description{
convert_BASF_FM
}
\examples{
 domainName<-'FM'
 domainData<-loadDomain(domainName)
 SEND<-convertBW(domainName, domainData)
 
}
