% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_FW.R
\name{convert_BASF_FW}
\alias{convert_BASF_FW}
\title{convert_BASFF_FW}
\usage{
convert_BASF_FW(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convert_BASFF_FW
}
\examples{
 domainName<-'FW'
 domainData<-loadDomain(domainName)
 SEND<-convertFW(domainName, domainData)

}
