% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_FX.R
\name{convert_BASF_FX}
\alias{convert_BASF_FX}
\title{convert_BASF_FX}
\usage{
convert_BASF_FX(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The dataframe containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe, use BASF Provantis Export files
}
\description{
convert_BASF_FX
}
\examples{
 domainName<-'Fx'
 domainData<-loadDomain(domainName)
 SEND<-convert_BASF_FX(domainName, domainData)
 
}
