% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_MA.R
\name{convert_BASF_MA}
\alias{convert_BASF_MA}
\title{convert_BASF_MA}
\usage{
convert_BASF_MA(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convert_BASF_MA
}
\examples{
 domainName<-'MA'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
