% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_OM.R
\name{convert_BASF_OM}
\alias{convert_BASF_OM}
\title{convert_BASF_OM
This module converts the Provantis Export for Organ Measurements to a SEND like structure leading to a OM Domain.}
\usage{
convert_BASF_OM(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convert_BASF_OM
This module converts the Provantis Export for Organ Measurements to a SEND like structure leading to a OM Domain.
}
\examples{
 domainName<-'OM'
 domainData<-loadDomain(domainName)
 SEND<-convert_BASF_OM(domainName, domainData)

}
