% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCL.R
\name{convertCL}
\alias{convertCL}
\title{convertCL}
\usage{
convertCL(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertCL
}
\examples{
 domainName<-'CL'
 domainData<-loadDomain(domainName)
 SEND<-convertCL(domainName, domainData)

}
