% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCV.R
\name{convertCV}
\alias{convertCV}
\title{convertCV
This module converts the Provantis Export for Cardiovascular Data to a SEND like structure leading to a CV Domain.}
\usage{
convertCV(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertCV
This module converts the Provantis Export for Cardiovascular Data to a SEND like structure leading to a CV Domain.
}
\examples{
 domainName<-'CV'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
