% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDM.R
\name{convertDM}
\alias{convertDM}
\title{convertDM
This module converts the Provantis Export for Demographics to a SEND like structure leading to a DM Domain.}
\usage{
convertDM(domainData)
}
\arguments{
\item{domainData}{The data frame containing the csv export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertDM
This module converts the Provantis Export for Demographics to a SEND like structure leading to a DM Domain.
}
\examples{
 SEND<-convertDM(domainData)

}
