% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDS.R
\name{convertDS}
\alias{convertDS}
\title{convertDS
This module converts the Provantis Export for Dispostion Data to a SEND like structure leading to a DS Domain.}
\usage{
convertDS(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertDS
This module converts the Provantis Export for Dispostion Data to a SEND like structure leading to a DS Domain.
}
\examples{
 domainName<-'DS'
 SEND<-convertDS(domainData)

}
