% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEG.R
\name{convertEG}
\alias{convertEG}
\title{convertEG}
\usage{
convertEG(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertEG
}
\examples{
 domainName<-'EG'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
