% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEX.R
\name{convertEX}
\alias{convertEX}
\title{convertEX}
\usage{
convertEX(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertEX
}
\examples{
 domainName<-'EX'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
