% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFW.R
\name{convertFW}
\alias{convertFW}
\title{convertFW}
\usage{
convertFW(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertFW
}
\examples{
 domainName<-'FW'
 domainData<-loadDomain(domainName)
 SEND<-convertFW(domainName, domainData)

}
