% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLB.R
\name{convertLB}
\alias{convertLB}
\title{convertLB}
\usage{
convertLB(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis for the LB parameters}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame for LB
}
\description{
convertLB
}
\examples{
 domainName<-'LB'
 SEND<-convertLB(domainData)

}
