% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMA.R
\name{convertMA}
\alias{convertMA}
\title{convertMA}
\usage{
convertMA(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertMA
}
\examples{
 domainName<-'MA'
 SEND<-convertMA(domainData)

}
