% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOM.R
\name{convertOM}
\alias{convertOM}
\title{convertOM
This module converts the Provantis Export for Organ Measurements to a SEND like structure leading to a OM Domain.}
\usage{
convertOM(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertOM
This module converts the Provantis Export for Organ Measurements to a SEND like structure leading to a OM Domain.
}
\examples{
 domainName<-'OM'
 SEND<-convertOM(domainData)

}
