% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTA.R
\name{convertTA}
\alias{convertTA}
\title{convertTA
This module converts the Provantis Export for Test Amount to a SEND like structure leading to a TA Domain.}
\usage{
convertTA(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertTA
This module converts the Provantis Export for Test Amount to a SEND like structure leading to a TA Domain.
}
\examples{
 domainName<-'TA'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
