% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTS.R
\name{convertTS}
\alias{convertTS}
\title{convertTS}
\usage{
convertTS(domainData)
}
\arguments{
\item{domainData}{The dataframe containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertTS
}
\examples{
 domainName<-'TS'
 domainData<-loadDomain(domainName)
 SEND<-convertTS(domainData)

}
