% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateDTC.R
\name{curateDTC}
\alias{curateDTC}
\title{curateDTC}
\usage{
curateDTC(d)
}
\arguments{
\item{d}{SEND domain data frame}
}
\value{
dd data frame for domain with DTC columns formatted as PosixCT
}
\description{
curateDTC
}
