% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{SENDIG dictionary}
\format{
\subsection{'csv'}{

A dataframe with 15 colums and 674 rows containg the SEND Domain Structure from the SEND Implementation Guide version 3.1
}
}
\source{
\url{https://www.cdisc.org/standards/foundational/send/sendig-v3-1-1}
}
\usage{
dictionary
}
\description{
SENDIG dictionary
}
\keyword{datasets}
