% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDomain.R
\name{exportDomain}
\alias{exportDomain}
\title{exportDomain}
\usage{
exportDomain(domain, fileformat, outdir)
}
\arguments{
\item{domain}{SEND Domain compiled by convert function}

\item{fileformat}{either csv, or pdf are supported}

\item{outdir}{directory name for saving the output files}
}
\description{
exportDomain
}
\examples{
domain<-'BW'
bw<-convertBW(domain)
exportDomain(bw, 'csv', 'outdir')

}
