% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDomain.R
\name{loadDomain}
\alias{loadDomain}
\title{loadDomain}
\usage{
loadDomain(domain, dir = dir, pattern = pattern, dec = dec)
}
\arguments{
\item{domain}{A character vector that contains the acronym for the domain}

\item{dir}{The directory name containng the input files}
}
\value{
data frame containing the data from the input file
}
\description{
loadDomain
}
\examples{
domain<-'BW'
bw<-loadDomain(domain, dir='/inputdir')
}
