% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSENDIG.R
\name{loadSENDIG}
\alias{loadSENDIG}
\title{loadSENDIG}
\usage{
loadSENDIG(domain = domain)
}
\arguments{
\item{domain}{SEND domain to be processed}
}
\value{
ddd Empty data frame for domain with column names extracted from SENDIG
}
\description{
loadSENDIG
}
\examples{
domain<-'BW'
d<-loadSENDIG(domain=domain)
}
