% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdSendig.R
\name{rdSendig}
\alias{rdSendig}
\title{rdSendig
This module reads the SENDIG}
\usage{
rdSendig(domainName)
}
\arguments{
\item{domain}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as dataframe
}
\description{
rdSendig
This module reads the SENDIG
}
\examples{
 domainName<-'BW'
 domainData<-rdSendig(domainName)

}
