% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_domain_manager.R
\docType{data}
\name{the}
\alias{the}
\title{Create a new environment and define a tibble with Acronyms and functions}
\format{
An object of class \code{environment} of length 1.
}
\usage{
the
}
\description{
Create a new environment and define a tibble with Acronyms and functions
}
\keyword{datasets}
