% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_BASF_CL.R
\name{convert_BASF_CL}
\alias{convert_BASF_CL}
\title{convert_BASF_CL}
\usage{
convert_BASF_CL(domainName, domainData)
}
\arguments{
\item{doCLinName}{A character vector that contains the acronym for the doCLin}

\item{doCLinData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted DoCLin as dataframe
}
\description{
convert_BASF_CL
}
\examples{
 doCLinName<-'CL'
 doCLinData<-loadDoCLin(domainName)
 SEND<-convertCL(domainName, domainData)

}
