% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBW.R
\name{convertBW}
\alias{convertBW}
\title{convertBW}
\usage{
convertBW(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The dataframe containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertBW
}
\examples{
 domainName<-'BW'
 domainData<-loadDomain(domainName)
 SEND<-convertBW(domainName, domainData)

}
