% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDM.R
\name{convertDM}
\alias{convertDM}
\title{convertDM
This module converts the Provantis Export for Demographics to a SEND like structure leading to a DM Domain.}
\usage{
convertDM(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertDM
This module converts the Provantis Export for Demographics to a SEND like structure leading to a DM Domain.
}
\examples{
 domainName<-'DM'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
