% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEG.R
\name{convertEG}
\alias{convertEG}
\title{convertEG}
\usage{
convertEG(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertEG
}
\examples{
 domainName<-'EG'
 SEND<-convertEG(domainData)

}
