% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMA.R
\name{convertMA}
\alias{convertMA}
\title{convertMA}
\usage{
convertMA(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as dataframe
}
\description{
convertMA
}
\examples{
 domainName<-'MA'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
