% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMI.R
\name{convertMI}
\alias{convertMI}
\title{convertMI}
\usage{
convertMI(domainName, domainData)
}
\arguments{
\item{domainName}{A character vector that contains the acronym for the domain}

\item{domainData}{The data frame containing the export of Provantis}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertMI
}
\examples{
 domainName<-'MI'
 domainData<-loadDomain(domainName)
 SEND<-convertMA(domainName, domainData)

}
