% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRE.R
\name{convertRE}
\alias{convertRE}
\title{convertRE
This module converts the Provantis Export for Respiratory Data to a SEND like structure leading to a RE Domain.}
\usage{
convertRE(domainData)
}
\arguments{
\item{domainData}{The data frame containing the export of Provantis}

\item{domainName}{A character vector that contains the acronym for the domain}
}
\value{
SEND Converted Domain as data frame
}
\description{
convertRE
This module converts the Provantis Export for Respiratory Data to a SEND like structure leading to a RE Domain.
}
\examples{
 domainName<-'RE'
 SEND<-convertMA(domainData)

}
