% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SendConv.R
\name{SendConv}
\alias{SendConv}
\title{SendConv}
\usage{
SendConv(domains, indir, outdir)
}
\arguments{
\item{domains}{character list of domains to be processed}

\item{indir}{directory for input files}

\item{outdir}{directory for outout files}
}
\description{
SendConv
}
\examples{
domains<-c('BW', 'LB')
indir<-'c:/temp/mquest_send/Provantis/DA/ARCH'
outdir<-'c:/temp/mquest_send/Provantis/DA/ARCH/output'
SendConv(domains, indir, outdir)
}
